DDL mapping

+++ miniprojekt/MiniDb-DDL/MiniDb/Payments.sql
@@ -2,10 +2,10 @@ create table Payments
 (
     payment_id     int auto_increment
         primary key,
-    reservation_id int      not null,
-    amount         int      not null,
-    date           datetime not null,
-    admin_id       int      null,
+    reservation_id int            not null,
+    amount         decimal(10, 2) not null,
+    date           datetime       not null,
+    admin_id       int            null,
     constraint Payments_Admins
         foreign key (admin_id) references Admins (admin_id),
     constraint Payments_Reservations
